Using custom archive
Basically, I created a custom archive for Station using xml files. Consistent usage of xml files is debatable, as databases read faster (in my experience) so I might switch to that later on. The only issue this has is when it reads binary files, so be sure to not include any