Updated chunking sort algorithm
The fault with using bubble sorting is that it ignores if the number in the last line in our list is the same number our current item starts with. It does not check if the current item number is greater than the last number. So I used my own sorting algorithm that works with respect to both the last known number, and the current number